XMLHttpRequest/tests/submissions/Opera/event-progress.htm

<!doctype html>
<html lang=en>
<meta charset=utf-8>
<title>XMLHttpRequest: progress event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#handler-xhr-onprogress" data-tested-assertations="/../.." />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#event-xhr-progress" data-tested-assertations="/../.." />
    <link rel="help" href="http://xhr.spec.whatwg.org/#the-send()-method" data-tested-assertations="/following::*//a[contains(@href, '#make-progress-notifications')]" />
    <link rel="help" href="http://xhr.spec.whatwg.org/#make-progress-notifications" data-tested-assertations="/.." />
    <link rel="help" href="http://xhr.spec.whatwg.org/#switch-done" data-tested-assertations="/following::li[5]" />
<div id="log"></div>
<script>
  var test = async_test();
  test.step(function() {
    //assert_unreached("TODO: finish test");
    var client = new XMLHttpRequest();
    client.onprogress = test.step_func(function() {
      // assert_equals(client.readyState, 3); // Once at 2, zero or more at 3?
      test.done();
    });
    client.onreadystatechange = test.step_func(function() {
      if (client.readyState === 4)
        assert_unreached("onprogress not called.");
    });
    client.open("GET", "resources/trickle.php");
    client.send(null);
  });
</script>
